refactor(worker): move builtin_agents.go → worker/config
BuiltinAgents relocates into the worker module (worker/config/). The
temporary cross-module import from worker/config to internal/config for
the AgentConfig type persists until Phase 4 consolidates the type.
