+++ worker/config/builtin_agents.go
@@ -1,6 +1,10 @@
 package config
 
-import "time"
+import (
+	"time"
+
+	internalconfig "github.com/Ivantseng123/agentdock/internal/config"
+)
 
 // BuiltinAgents is the canonical registry of agent CLI configurations shipped
 // with AgentDock. Config files may override individual entries by defining a
@@ -8,7 +12,11 @@ import "time"
 //
 // Adding a new built-in agent: just add an entry here. Existing users get it
 // automatically on next startup; no `agentdock init` rerun needed.
-var BuiltinAgents = map[string]AgentConfig{
+//
+// Note: AgentConfig still lives in internal/config. Phase 4 will consolidate
+// the type here or in shared — until then, this file uses internalconfig.AgentConfig
+// to avoid duplicating the type definition.
+var BuiltinAgents = map[string]internalconfig.AgentConfig{
 	"claude": {
 		Command:  "claude",
 		Args:     []string{"--print", "--output-format", "stream-json", "-p", "{prompt}"},
